+++ tests/test_doublyrobust.py
@@ -54,7 +54,7 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.psi, r_rd)
+        npt.assert_allclose(tmle.risk_difference, r_rd)
 
     def test_match_r_tmle_rd_ci(self, df):
         r_ci = -0.1541104, -0.01470202
@@ -63,7 +63,7 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.confint, r_ci, rtol=1e-5)
+        npt.assert_allclose(tmle.risk_difference_ci, r_ci, rtol=1e-5)
 
     def test_match_r_tmle_riskratio(self, df):
         r_rr = 0.5344266
@@ -72,7 +72,7 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.psi, r_rr)
+        npt.assert_allclose(tmle.risk_ratio, r_rr)
 
     def test_match_r_tmle_rr_ci(self, df):
         r_ci = 0.2773936, 1.0296262
@@ -81,16 +81,16 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.confint, r_ci, rtol=1e-5)
+        npt.assert_allclose(tmle.risk_ratio_ci, r_ci, rtol=1e-5)
 
     def test_match_r_tmle_oddsratio(self, df):
-        r_rr = 0.4844782
+        r_or = 0.4844782
         tmle = TMLE(df, exposure='art', outcome='dead', measure='odds_ratio')
         tmle.exposure_model('male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0', print_results=False)
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.psi, r_rr)
+        npt.assert_allclose(tmle.odds_ratio, r_or)
 
     def test_match_r_tmle_or_ci(self, df):
         r_ci = 0.232966, 1.007525
@@ -99,7 +99,7 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.confint, r_ci, rtol=1e-5)
+        npt.assert_allclose(tmle.odds_ratio_ci, r_ci, rtol=1e-5)
 
     def test_symmetric_bounds_on_gW(self, df):
         r_rd = -0.08203143
@@ -110,8 +110,8 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.psi, r_rd)
-        npt.assert_allclose(tmle.confint, r_ci, rtol=1e-5)
+        npt.assert_allclose(tmle.risk_difference, r_rd)
+        npt.assert_allclose(tmle.risk_difference_ci, r_ci, rtol=1e-5)
 
     def test_asymmetric_bounds_on_gW(self, df):
         r_rd = -0.08433208
@@ -122,8 +122,8 @@ class TestTMLE:
         tmle.outcome_model('art + male + age0 + age_rs1 + age_rs2 + cd40 + cd4_rs1 + cd4_rs2 + dvl0',
                            print_results=False)
         tmle.fit()
-        npt.assert_allclose(tmle.psi, r_rd)
-        npt.assert_allclose(tmle.confint, r_ci, rtol=1e-5)
+        npt.assert_allclose(tmle.risk_difference, r_rd)
+        npt.assert_allclose(tmle.risk_difference_ci, r_ci, rtol=1e-5)
 
     def test_sklearn_in_tmle(self, df):
         log = LogisticRegression(penalty='l1', C=1.0, random_state=201)
@@ -134,7 +134,7 @@ class TestTMLE:
         # Dropping since Linux RNG does not match my OS (windows)
         # npt.assert_allclose(tmle.psi, -0.07507877527854623)
         # npt.assert_allclose(tmle.confint, [-0.15278930211034644, 0.002631751553253986], rtol=1e-5)
-        # Test now only checks no errors are thrown. To fix later
+        # TODO Test now only checks no errors are thrown. To fix later
 
 
 class TestAIPTW: